refactor(worldgen): extract gate greeter NPC logic to worldgen/town/npcs_bootstrap.js and wire imports in town_gen.js
